quiet a couple sctp warnings
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@4989 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/sofia-sip/libsofia-sip-ua/tport/tport_type_sctp.c
@@ -96,8 +96,8 @@ static int tport_sctp_init_socket(tport_primary_t *pri,
 				  int socket,
 				  char const **return_reason);
 static int tport_recv_sctp(tport_t *self);
-static int tport_send_sctp(tport_t const *self, msg_t *msg,
-			   msg_iovec_t iov[], int iovused);
+static ssize_t tport_send_sctp(tport_t const *self, msg_t *msg,
+			       msg_iovec_t iov[], size_t iovused);
 
 tport_vtable_t const tport_sctp_client_vtable =
 {
@@ -254,8 +254,8 @@ int tport_recv_sctp(tport_t *self)
   return 2;
 }
 
-static int tport_send_sctp(tport_t const *self, msg_t *msg,
-			   msg_iovec_t iov[], int iovused)
+static ssize_t tport_send_sctp(tport_t const *self, msg_t *msg,
+			       msg_iovec_t iov[], size_t iovused)
 {
